Boltz/ConnectionIF.hpp: Create generic interface to connection.
